% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entropy.R
\name{entropy}
\alias{entropy}
\title{Compute entropy of a categorical variable}
\usage{
entropy(x, normalise = TRUE)
}
\arguments{
\item{x}{A character vector representing observed categorical values.}

\item{normalise}{Logical. If \code{TRUE}, entropy is normalised to range from 0 to 1. If \code{FALSE}, returns raw Shannon entropy.}
}
\value{
A numeric value representing the entropy of \code{x}.
}
\description{
Computes the entropy of a categorical variable, with optional normalisation.
}
\keyword{internal}
