% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_alpha_by_kappa.R
\name{filter_alpha_by_kappa}
\alias{filter_alpha_by_kappa}
\title{Filter ancestries by generation threshold}
\usage{
filter_alpha_by_kappa(out, kappa_max)
}
\arguments{
\item{out}{A data frame of class \code{outbreaker_chains}.}

\item{kappa_max}{A single integer specifying the maximum allowed number of generations (\code{kappa}).}
}
\value{
A modified \code{outbreaker_chains} data frame with alpha values filtered based on the kappa threshold.
}
\description{
Filters \code{alpha} values in an \code{outbreaker_chains} object based on a specified threshold for \code{kappa}, removing ancestors beyond the given number of generations.
}
\examples{
\dontrun{
filter_alpha_by_kappa(out, kappa_max = 1)
}

}
