% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accuracy.R
\name{get_accuracy}
\alias{get_accuracy}
\title{Calculate accuracy of a outbreak reconstruction
Accuracy of outbreak reconstruction is defined as the proportion of correctly assigned ancestries
across the posterior chain.}
\usage{
get_accuracy(chain, true_tt)
}
\arguments{
\item{chain}{A list of data frames, each containing a posterior chain of transmission trees.}

\item{true_tt}{A data frame containing the true transmission tree. Columns 'from' and 'to' are required.}
}
\value{
A numeric vector of accuracy values for each posterior tree.
}
\description{
Calculate accuracy of a outbreak reconstruction
Accuracy of outbreak reconstruction is defined as the proportion of correctly assigned ancestries
across the posterior chain.
}
