% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trees.R
\name{get_trees}
\alias{get_trees}
\title{Extract posterior transmission trees}
\usage{
get_trees(out, ids, kappa = FALSE, t_inf = FALSE, ...)
}
\arguments{
\item{out}{A data frame of class \code{outbreaker_chains}.}

\item{ids}{A character or integer vector of case IDs from the original linelist.}

\item{kappa}{A logical indicating whether to include \code{kappa} values in the output. Default is \code{FALSE}.}

\item{t_inf}{A logical indicating whether to include infection times (\code{t_inf}) in the output. Default is \code{FALSE}.}

\item{...}{Additional named vectors from the original linelist to include as columns (e.g., \code{loc = linelist$location}).}
}
\value{
A list of data frames. Each data frame contains 'from' and 'to' columns, and optionally includes kappa, t_inf, and additional columns.
}
\description{
Generates a list of data frames representing posterior transmission trees from an \code{outbreaker_chains} object.
Each tree is represented as a data frame with 'from' and 'to' columns, and optionally includes additional columns.
}
\examples{
\dontrun{
get_trees(out, ids = ids, group = group, dates = dates)
}

}
