% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify}
\alias{identify}
\title{Replace integers in outbreaker2 output with unique identifiers}
\usage{
identify(out, ids)
}
\arguments{
\item{out}{A data frame of class \code{outbreaker_chains}.}

\item{ids}{A vector of IDs from the original linelist (see \code{outbreaker2::outbreaker_data()}).}
}
\value{
A data frame of class \code{outbreaker_chains} with integers replaced by the corresponding IDs.
}
\description{
Replace integers in outbreaker2 output with unique identifiers
}
\examples{
identify(out, id = linelist$name)

}
