% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_linelist}
\alias{is_linelist}
\title{Check if the object conforms to linelist expectations}
\usage{
is_linelist(linelist)
}
\arguments{
\item{linelist}{The object to be checked.}
}
\value{
TRUE if the object is a data frame with an 'id' column; otherwise, stops with an error.
}
\description{
This function checks if the input is a data frame and contains an 'id' column,
as expected for a linelist.
}
\keyword{internal}
