% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outbreaker_chains.R
\name{is_outbreaker_chains}
\alias{is_outbreaker_chains}
\title{Check if an object is a data frame of class 'outbreaker_chains'}
\usage{
is_outbreaker_chains(out)
}
\arguments{
\item{out}{The object to be checked.}
}
\value{
TRUE if the object is a data frame and of class 'outbreaker_chains', otherwise stops with an error.
}
\description{
This function checks if the input is a data frame and of class 'outbreaker_chains'.
}
\keyword{internal}
