% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linelist}
\alias{linelist}
\title{Simulated linelist with group labels}
\format{
A data frame with 30 rows and 5 columns:
\describe{
\item{id}{Case ID}
\item{name}{Simulated first name}
\item{group}{Group label: \code{"patient"} or \code{"hcw"}}
\item{onset}{Date of symptom onset}
\item{sample}{Date of sample collection}
}
}
\usage{
linelist
}
\description{
A simulated linelist derived from \code{fake_outbreak}, where cases are assigned to the patient or hcw group.
First names are randomly generated using the \pkg{randomNames} package.
}
\examples{
head(linelist)
}
\seealso{
\code{\link[outbreaker2]{fake_outbreak}}
}
\keyword{datasets}
