% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{out}
\alias{out}
\title{outbreaker2 toy dataset}
\format{
An \code{outbreaker_chains} object.
}
\source{
\url{https://www.repidemicsconsortium.org/outbreaker2/articles/introduction.html}
}
\usage{
out
}
\description{
The outbreaker2 result generated from the example in the
\href{https://www.repidemicsconsortium.org/outbreaker2/articles/introduction.html}{outbreaker2 vignette}.
This dataset was produced by running \code{outbreaker()} on the \code{fake_outbreak} data.
}
\keyword{datasets}
