% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.outbreaker_chains.R
\name{sample.outbreaker_chains}
\alias{sample.outbreaker_chains}
\title{Sample rows from an outbreaker_chains object}
\usage{
sample.outbreaker_chains(x, ...)
}
\arguments{
\item{x}{An object of class \code{outbreaker_chains}, which also inherits from \code{data.frame}.}

\item{...}{Additional arguments to be passed to \code{sample()}, such as \code{size} or \code{replace}.}
}
\value{
An object of class \code{outbreaker_chains}, with sampled rows.
}
\description{
This function samples rows from an object of class \code{outbreaker_chains}.
}
