% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entropy.R
\name{entropy}
\alias{entropy}
\title{Compute entropy of a categorical variable}
\usage{
entropy(x)
}
\arguments{
\item{x}{A character vector representing observed categorical values.}
}
\value{
A numeric value representing the entropy of \code{x}.
}
\description{
Computes the entropy of a categorical variable based on its observed probabilities.
}
\keyword{internal}
