% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_alpha_by_kappa .R
\name{filter_alpha_by_kappa}
\alias{filter_alpha_by_kappa}
\title{Filter alpha values based on kappa threshold}
\usage{
filter_alpha_by_kappa(out, kappa_max)
}
\arguments{
\item{out}{The outbreaker2 data.frame.}

\item{kappa_max}{The threshold value for kappa.}
}
\value{
The modified outbreaker2 data.frame with filtered alpha values.

#' @examples
\dontrun{
filter_alpha_by_kappa(out, kappa_max = 1)
}
}
\description{
This function filters the alpha values in an outbreaker2 data.frame based on a given threshold for kappa. i.e. it keeps only the ancestors that are below a given number of generations.
}
