% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accuracy.R
\name{get_accuracy}
\alias{get_accuracy}
\title{Calculate accuracy of a outbreak reconstruction}
\usage{
get_accuracy(est_tt, true_tt)
}
\arguments{
\item{est_tt}{A list of transmission trees. See \code{get_trees}.
Each element is a data frame containing the transmission tree with columns 'from' and 'to'.}

\item{true_tt}{A data frame containing the true transmission tree. Columns 'from' and 'to' are required.}
}
\value{
A numeric vector of accuracy values for each posterior tree.
}
\description{
Accuracy of outbreak reconstruction is defined as the proportion of correctly assigned ancestries
across the posterior sample.
}
