% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trees.R
\name{get_trees}
\alias{get_trees}
\title{Return a list of posterior transmission trees from an outbreaker2 object.}
\usage{
get_trees(out, ids, kappa = FALSE, t_inf = FALSE, ...)
}
\arguments{
\item{out}{A data frame of class \code{outbreaker_chains}.}

\item{ids}{A vector of IDs from the original linelist.}

\item{kappa}{A logical value indicating whether to include the kappa values in the output.}

\item{t_inf}{A logical value indicating whether to include the kappa values in the output.}

\item{...}{Additional columns from the original linelist to include (e.g. loc = linelist$location).}
}
\value{
A list of data frames. Each data frame has 'from' and 'to' columns,
and additional columns based on the additional arguments.
}
\description{
This function takes an \code{outbreaker_chains} object and returns a list of data frames with the 'from' and 'to' columns, and additional columns if provided.
The additional arguments should be vectors of values, and the name of the argument will be used as the name of the additional column.
The additional columns will be named 'from_' and 'to_' followed by the name of the argument.
}
\examples{
\dontrun{
get_trees(out, ids = ids, group = group, dates = dates)
}

}
