% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttable.R
\name{ttable}
\alias{ttable}
\title{Transmission table}
\usage{
ttable(from, to, levels = NULL, ...)
}
\arguments{
\item{from}{A vector of infectors.}

\item{to}{A vector of infectees.}

\item{levels}{Optional. A vector of levels for the factors. Defaults to unique, sorted values of 'from' and 'to'.}

\item{...}{Additional arguments to be passed to the 'table' function.}
}
\value{
A contingency table representing the counts of infectors (rows) and infectees (columns) in each group (level).
}
\description{
Generates a contingency table given the 'from' and 'to' vectors.
}
\examples{
from <- c("A", "A", NA, "C", "C", "C")
to <- c("A", "B", "B", "C", "C", "C")
ttable(from, to)
}
